solution/requirements.txt
psycopg2-binary
pandas
requests
gspread
google-auth
openpyxl
google-api-python-client
pytest